Fix crash in 64-bit mode
The problem was caused by using the (legacy) float data type instead of
CGFloat for the color components. This caused the alphaComponent in
setColor to be 0, which in turn raised this assertion error:

  NSAssert((alphaComponent > 0.999), @"color must be opaque");

+++ SharedCode/AMIndeterminateProgressIndicatorCell.m
@@ -50,7 +50,7 @@
 
 - (void)setColor:(NSColor *)value
 {
-	float alphaComponent;
+	CGFloat alphaComponent;
 	if (color != value) {
 		[color release];
 		color = [value retain];
